fix: moves brownie.utils.color to brownie._color
