+++ brownie/_cli/accounts.py
@@ -6,9 +6,9 @@ import sys
 from pathlib import Path
 
 from brownie import accounts
+from brownie._colors import color, notify
 from brownie._config import _get_data_folder
 from brownie.convert import to_address
-from brownie.utils import color, notify
 from brownie.utils.docopt import docopt
 
 __doc__ = """Usage: brownie accounts <command> [<arguments> ...] [options]
