+++ brownie/_cli/analyze.py
@@ -13,6 +13,7 @@ from pythx import Client, ValidationError
 from pythx.middleware import ClientToolNameMiddleware, GroupDataMiddleware
 
 from brownie import project
+from brownie._colors import color, notify
 from brownie._config import (
     CONFIG,
     __version__,
@@ -20,7 +21,6 @@ from brownie._config import (
     _update_argv_from_docopt,
 )
 from brownie.exceptions import ProjectNotFound
-from brownie.utils import color, notify
 from brownie.utils.docopt import docopt
 
 __doc__ = """Usage: brownie analyze [options] [--async | --interval=<sec>]
