+++ brownie/_cli/ethpm.py
@@ -7,10 +7,10 @@ import sys
 import yaml
 
 from brownie import accounts, network
+from brownie._colors import color, notify
 from brownie._config import BROWNIE_FOLDER, _get_data_folder
 from brownie.exceptions import ProjectNotFound, UnknownAccount
 from brownie.project import check_for_project, ethpm
-from brownie.utils import color, notify
 from brownie.utils.docopt import docopt
 
 __doc__ = """Usage: brownie ethpm <command> [<arguments> ...] [options]
