--- brownie/utils/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python3
-
-from .color import Color, notify  # noqa 401
-
-color = Color()
